+++ R/accessors.spdur.R
@@ -1,68 +1,28 @@
-#' AIC method for spdur
+#' Accessor methods for spdur Objects
 #' 
-#' Computes the Akaike Information Criterion for an \code{spdur} class object.
+#' Several standard accessor methods for a \code{spdur} class object.
 #' 
-#' @method AIC spdur
+#' @param object an object inheriting from class \code{spdur}.
+#' @param x  \code{spdur} class object for terms
+#' @param model return full model, or only duration or risk equations, or 
+#'   distribution parameters.
+#' @param \dots not used
 #' 
-#' @param object An object of class \code{spdur}.
-#' @param \dots Optional arguments.
-#' @param k The penalty parameter, by default 2. For \code{\link{BIC.spdur}},
-#' the penalty parameter equals \code{log(N)}.
+#' @seealso \code{\link{AIC.spdur}}, \code{\link{BIC.spdur}}
 #' 
-#' @seealso \code{link{AIC}}, \code{link{BIC.spdur}}
-#' 
-#' @examples
-#' data(model.coups)
-#' AIC(model.coups)
-#'
-#' @export 
-#' @importFrom stats AIC
-AIC.spdur <- function(object, ..., k = 2)
-{
-  npar <- length(object$coefficients)
-  lnL <- logLik(object)
-  aic <- as.vector(-2*lnL + k * npar)
-  
-  return(aic)
-}
-
-#' BIC method for spdur
-#' 
-#' Computes the Bayesian Information Criterion for an \code{spdur} class object.
-#' 
-#' @method BIC spdur
-#' 
-#' @param object An object of class \code{spdur}.
-#' @param \dots Optional arguments.
-#' 
-#' @details Computed as \code{AIC(object, k = log(nobs(object)))}.
-#' 
-#' @seealso \code{\link{BIC}}, \code{\link{AIC.spdur}}
+#' @name accessors
 #' 
 #' @examples
 #' data(model.coups)
-#' BIC(model.coups)
 #' 
-#' @export
-#' @importFrom stats BIC
-BIC.spdur <- function(object, ...)
-{
-  bic <- AIC(object, k = log(nobs(object)))
-  
-  return(bic)
-}
+NULL
+#> NULL
 
-#' Log-Likelihood of an spdur Object
-#' 
-#' Returns the log-likelihood of a \code{spdur} class object.
+#' @rdname accessors
 #' 
 #' @method logLik spdur
 #' 
-#' @param object an object inheriting from class \code{spdur}.
-#' @param \dots not used
-#' 
-#' @examples
-#' data(model.coups)
+#' @examples 
 #' logLik(model.coups)
 #' 
 #' @export
@@ -78,14 +38,10 @@ logLik.spdur <- function(object, ...)
   return(val)
 }
 
-#' Number of Observations in a spdur Object
-#' 
-#' Extract the number of observations in a \code{spdur} class model object.
-#' 
-#' @param object an object inheriting from class \code{spdur}.
-#' @param \dots not used.
-#' 
-#' @seealso \code{\link{AIC.spdur}}, \code{\link{BIC.spdur}}
+#' @rdname accessors
+#' @method nobs spdur
+#' @examples 
+#' nobs(model.coups)
 #' 
 #' @export
 #' @importFrom stats nobs
@@ -93,4 +49,91 @@ nobs.spdur <- function(object, ...)
 {
   val <- object$obs
   return(val)
+}
+
+#' @rdname accessors
+#' @method coef spdur
+#' @examples 
+#' coef(model.coups)
+#' 
+#' @export
+#' @importFrom stats coef
+coef.spdur <- function (object, model = c("full", "duration", "risk", "distr"), ...) 
+{
+  model <- match.arg(model)
+  res  <- object$coefficients
+  res  <- switch(
+    model, 
+    full = structure(c(res$duration, res$risk, res$distr), 
+      .Names = c(paste("duration", names(res$duration), sep = "_"), 
+                 paste("risk", names(res$risk), sep = "_"),
+                 names(res$distr))), 
+    duration = res$duration, 
+    risk = res$risk,
+    distr = res$distr)
+  return(res)
+}
+
+#' @rdname accessors
+#' @method vcov spdur
+#' @examples 
+#' vcov(model.coups)
+#' 
+#' @export
+#' @importFrom stats vcov
+vcov.spdur <- function (object, model = c("full", "duration", "risk", "distr"), ...) 
+{
+  model <- match.arg(model)
+  res   <- object$vcov
+  if (model == "full") 
+    return(res)
+  
+  # otherwise:
+  cf <- object$coefficients[[model]]
+  wi <- seq(along = object$coefficients$duration)
+  di <- nrow(res)
+  res <- if (model == "duration") {
+    res[wi, wi]
+  } else if (model == "risk") {
+    res <- res[-c(wi, di), -c(wi, di)]
+  } else {
+    res <- res[di, di, drop = FALSE]
+  }
+  colnames(res) <- rownames(res) <- names(cf)
+  return(res)
+}
+
+#' @rdname accessors
+#' @method model.matrix spdur
+#' @examples 
+#' head(model.matrix(model.coups))
+#' 
+#' @export
+#' @importFrom stats model.matrix
+model.matrix.spdur <- function(object, model = c("duration", "risk"), ...) 
+{
+  model <- match.arg(model)
+  res  <- switch(
+    model, 
+    duration = object$mf.dur, 
+    risk = object$mf.risk)
+  return(res)
+}
+
+#' @rdname accessors
+#' @method terms spdur
+#' @examples 
+#' terms(model.coups)
+#' 
+#' @export
+#' @importFrom stats terms
+terms.spdur <- function(x, model = c("duration", "risk"), ...) 
+{
+  model <- match.arg(model)
+  res  <- switch(
+    model, 
+    duration = terms(x$mf.dur), 
+    risk = terms(x$mf.risk)
+  )
+  return(res)
 }